requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-cloud-storage
firebase
firebase-admin
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
ffmpeg-python==0.2.0
newsapi-python==0.2.6
Pillow==9.1.1
python-dotenv==0.20.0
replicate==0.0.1a11
requests==2.27.1
urllib3==1.26.9
